[gdb/macros] Handle 64-bit dwarf in gdb.base/fission-macro.exp
When enabling 64-bit dwarf in the test-case gdb.base/fission-macro.exp, we run
into:
...
(gdb) break -qualified main^M
DW_MACRO_define_strx pointing outside of .debug_str.dwo section \
  [in module fission-macro-5-640.dwo]^M
(gdb) FAIL: gdb.base/fission-macro.exp: dwarf_version=5: dwarf_bits=64: \
  gdb_breakpoint: set breakpoint at main
...

A dwarf-5 .debug_str_offsets section starts with a header consisting of:
- an initial length (4 bytes for 32-bit and 12 bytes for 64-bit),
- a 2 byte version string, and
- 2 bytes padding
so in total 8 bytes for 32-bit and 16 bytes for 64-bit.

This offset is calculated here in dwarf_decode_macros:
...
      str_offsets_base = cu->header.addr_size;
...
and in both cases this evaluates to 8.

So in the 64-bit case, we interpret the last 8 bytes of the header as an offset.

Fix this by computing str_offsets_base correctly for dwarf-5, for both the
32-bit and 64-bit case.

Tested on x86_64-linux.

Tested test-case using a current gcc trunk build, and gcc 14.