[gdb/macros] Work around a gcc PR fixed in gcc 9
When running test-case gdb.base/fission-macro-i.exp with gcc 8, we run into:
...
(gdb) set complaints 0^M
Offset from DW_FORM_GNU_str_index or DW_FORM_strx pointing outside of \
  .debug_str.dwo section in CU at offset 0x0 \
  [in module fission-macro-i-v5-b32-s0]^M
(gdb) FAIL: gdb.base/fission-macro-i.exp: dwarf_version=5: dwarf_bits=32: \
  strict_dwarf=0: set complaints 0
...

We're using dwarf5, so the .debug_str_offsets section is supposed to have a
header, but there's none:
...
        .section        .debug_str_offsets.dwo,"e",@progbits
        .long   0       # indexed string 0x0: GNU C17 8.2.1 20180831 ...
        .long   0x87    # indexed string 0x1: FIRST 1
        .long   0x8f    # indexed string 0x2: /home/vries/gdb
        .long   0x9f    # indexed string 0x3: main
        .long   0xa4    # indexed string 0x4: fission-macro-i.i
        .long   0xe1    # indexed string 0x5: SECOND 2
...

In other words, we have a v4-style .debug_str_offsets section.

Workaround this by detecting that it's not a valid v5 header by comparing the
initial length to the the section size, and handling it as a v4 section.

IWBN to enable this workaround only for gcc version < 9, but unfortunately the
workaround is required to correctly read the producer string, so that's not
possible.

Tested on x86_64-linux.

Tested test-case (and gdb.base/fission-macro.exp) using a current gcc trunk
build and gcc 7-14.