[gdb/testsuite] Make test-case gdb.base/watchpoint-running.exp more reliable
Test-case gdb.base/watchpoint-running.exp does the following:
...
    set val1 ""
    gdb_test_multiple "print global_var" "global_var once" {
	-re -wrap " = ($::decimal)" {
	    set val1 $expect_out(1,string)
	    pass "$gdb_test_name"
	}
    }

    # Sleep for a bit, so the variable is sure to be incremented, if
    # indeed we managed to get the target running.
    sleep 1

    set val2 ""
    gdb_test_multiple "print global_var" "global_var twice" {
	-re -wrap " = ($::decimal)" {
	    set val2 $expect_out(1,string)
	    pass "$gdb_test_name"
	}
    }
...

The context is that the inferior is running in the background, incrementing
global_var every 0.1s.

The "sleep 1" is a trade off between these two issues:
- the normal period to wait for things to happen in the testsuite is $timeout
  (by default 10 seconds), but waiting for that amount of time will usually
  mean unnecessarily slowing down the test-case significantly.
- reading global_var twice without a delay inbetween has a significant chance
  of resulting in the same value.

This is both unreliable (the sleep 1 may not be long enough) and possibly too
slow (the sleep 1 may longer than necessary).

Fix this by doing the second read of global_var in a loop.

The loop times out after $timeout, restoring the usual timeout period and
making the test more reliable, while only waiting if necessary.

Furthermore:
- move the first read of global_var to before the "continue&", and
- move the usleep in the .c file to after the increment,
allowing the second read to get the value of the first increment without any
waiting.

Tested on x86_64-linux.