[gdb/testsuite] Remove hardcoded timeout in gdb.base/watchpoint-running.exp
In test-case gdb.base/watchpoint-running.exp we have:
...
    # Check that the watchpoint triggers.

    save_vars ::timeout {
	if {!$hw} {
	    # This doesn't currently work with software watchpoints.
	    # GDB should transparently temporarily pause the inferior,
	    # to force it to single step, but it doesn't, so the
	    # thread continues running free.
	    setup_kfail gdb/31833 *-*-*

	    # No point waiting too much for output we know isn't
	    # coming.
	    set ::timeout 1
	}
    }
...

The intention here is that if somebody adds the missing functionality and runs
the testsuite, the kfail will turn into a kpass, notifiying the author that
this test-case needs updating, and might be able to function as regression
test.

However, hardcoding a timeout of 1 makes this unreliable.  Software
watchpoints are slow, and without having implemented the functionality and
running the test-case, it's hard to predict whether the timeout is sufficient
to produce a kpass.

This could be fixed by adding a loop that:
- checks the value of global_var, and
- terminates when the value has been increment (so the watchpoint should have
  triggered), and
- terminates when the regular $timeout.

However, I think this adds too much complexity for this non-essential feature
of the test-case.

Fix this by just producing a kfail.

Tested on x86_64-linux.