[gdb/tdep] Use cached breakpoint kind in breakpoint_kind
With test-case gdb.base/solib-probes-nosharedlibrary.exp on arm-linux, I get:
...
(gdb) continue^M
Continuing.^M
^M
Program received signal SIGILL, Illegal instruction.^M
0xf7fd267a in ?? ()^M
(gdb) FAIL: $exp: continue to breakpoint: main
...

This happens as follows:
- the starti command is called,
- gdb loads symbols for ld.so,
- the "shlib events" breakpoint is inserted at the address of _dl_debug_state,
- execution stops at ld.so _start, and the inserted breakpoint is removed,
- the nosharedlibrary command is called,
- gdb forgets about all shared library symbols,
- the continue command is called,
- gdb re-inserts the "shlib events" breakpoint, but the call to
  breakpoint_kind returns the wrong value: ARM_BP_KIND_ARM instead of
  ARM_BP_KIND_THUMB, so the wrong type of breakpoint is inserted, and
- when execution hits the breakpoint, the SIGILL triggers.

In order to correctly set a breakpoint on an address for arm, gdb needs to
known whether that address is in thumb or arm code, and the problem is that
that information is no longer available due to the nosharedlibrary command.

However, the breakpoint kind is is available from the previous time the
breakpoint was inserted.

Fix this in breakpoint_kind, by using the info cached in bl->target_info.kind,
if available.

Tested on arm-linux.

PR tdep/31817
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=31817