[gdb/symtab] Don't complain about stdc-predef.h include
With a minimal test-case test.c:
...
int main (void) { return 0; }
...
compiled like this:
...
$ gcc test.c -g3 -gdwarf-5
...
I get:
...
$ gdb -q -batch -iex "set complaints 100" a.out -ex "p main"
During symbol reading: debug info gives source 2 included from file at zero line 0
$1 = {int (void)} 0x4101ac <main>
...

The complaint is about the second start_file here:
...
Contents of the .debug_macro section:

  Offset:                      0
  Version:                     5
  Offset size:                 4
  Offset into .debug_line:     0

 DW_MACRO_import - offset : 0x1a
 DW_MACRO_start_file - lineno: 0 filenum: 1
 DW_MACRO_start_file - lineno: 0 filenum: 2
 DW_MACRO_import - offset : 0xb1c
 DW_MACRO_end_file
 DW_MACRO_end_file
...
where filenum 2 refers to /usr/include/stdc-predef:
...
 The Directory Table
  Entry Name
  0     /home/vries/gdb
  1     /usr/include

 The File Name Table
  Entry Dir     Name
  0     0       test.c
  1     0       test.c
  2     1       stdc-predef.h
...

The "DW_MACRO_start_file - lineno: 0 filenum: 2" entry represents an implicit
include of /usr/include/stdc-predef, so the line number of 0 is correct.

But the complaints mechanism expects entries with line number zero to occur
only before the first start_file, which is the entry for "the source file
submitted to the compiler for compilation".

And moving the "DW_MACRO_start_file - lineno: 0 filenum: 2" entry to before
the first start_file doesn't work, because it would make it the
first start_file.

The standard requires that:
- the first start_file entry has lineno 0, and
- all define and undefine entries representing compilation options:
  - appear before the first start_file entry, and
  - have lineno 0.

Both these conditions are met, so the generated dwarf is standard compliant.

OTOH, the complaint is useful because it's likely to catch cases of incorrect
dwarf assembly.

So, fix this by simply not complaining about this corner-case.

Tested on aarch64-linux.