[gdb/tdep] Use gdbarch_breakpoint_kind_from_pc_v2 for arm
Rewrite arm_breakpoint_kind_from_pc into the gdbarch_breakpoint_kind_from_pc_v2
style, allowing to cache thumbness of the breakpoint address for a subsequent
call, and use the v2 variant (indirectly) in code_breakpoint::insert_location.

This happens to fix the following failure with test-case
gdb.base/solib-probes-nosharedlibrary.exp on arm-linux:
...
(gdb) continue^M
Continuing.^M
^M
Program received signal SIGILL, Illegal instruction.^M
0xf7fd267a in ?? ()^M
(gdb) FAIL: $exp: continue to breakpoint: main
...

This failure happens as follows:
- the starti command is called,
- gdb loads symbols for ld.so,
- the "shlib events" breakpoint is inserted at the address of _dl_debug_state,
- execution stops at ld.so _start, and the inserted breakpoint is removed,
- the nosharedlibrary command is called,
- gdb forgets about all shared library symbols,
- the continue command is called,
- gdb re-inserts the "shlib events" breakpoint, but the call to
  breakpoint_kind returns the wrong value: ARM_BP_KIND_ARM instead of
  ARM_BP_KIND_THUMB, so the wrong type of breakpoint is inserted, and
- when execution hits the breakpoint, the SIGILL triggers.

In order to correctly set a breakpoint on an address for arm, gdb needs to
known whether that address is in thumb or arm code, and the problem is that
that information is no longer available due to the nosharedlibrary command.

But because we now cache the thumbness of the breakpoint address alongside the
breakpoint address, that's no longer a problem.

Tested on arm-linux.

PR tdep/31817
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=31817