[gdb/testsuite] Improve .debug_names dwarf assembler
We have a few dwarf assembly test-cases that generate a .debug_names section.

When comparing that to a .debug_names section as generated by gdb-add-index,
the DW_IDX_die_offset and DW_IDX_GNU_language attributes are missing.

Improve on this by adding the DW_IDX_GNU_language attribute in those
test-cases.

While we're at it, use .uleb128 instead of .byte in a few places.

Tested on x86_64-linux.