[gdb/contrib] Add replace-on-error.sh
Add a refactoring script gdb/contrib/replace-on-error.sh, that can be used to
rename things in the sources, driven by errors or warnings in the build log:
...
$ ./src/gdb/contrib/replace-on-error.sh \
  "error:.*::raw_collect(.* is deprecated \[-Werror=deprecated-declarations\]" \
  raw_supply \
  deprecated_raw_colllect \
  build-log.txt
...

A gotcha of the current implementation is that when replacing in a macro (as
here in ia64-linux-nat.c):
...
 #define COPY_REG(_idx_,_regi_) \
   if ((regno == -1) || regno == _regi_) \
     regcache->raw_collect (_regi_, regp + _idx_)
...
it'll trigger multiple times and we end up with something like
regcache->deprecateddeprecated...deprecatedeprecated_raw_collect.

Tested on x86_64-linux.  Verified with shellcheck.