[gdb/build] Deprecate deprecated reg_buffer::raw_collect
We have in gdb/regcache.h:
...
  void raw_collect (int regnum, gdb::array_view<gdb_byte> dst) const override;

  /* Deprecated overload of the above.  */
  void raw_collect (int regnum, void *dst) const;
...

The deprecation is only mentioned in the commment, so it's easy to miss both
that:
- a new use of raw_collect is a deprecated one, and
- an existing use of raw_collect is a deprecated one.

Tagging the deprecated raw_collect with [[deprecated]] will fix the former,
but also trigger -Werror=deprecated-declarations errors for the latter.

Fix this by:
- adding deprecated_raw_collect, a wrapper for the deprecated raw_collect
  that silences -Wdeprecated-declarations,
- tagging the deprecated raw_collect with [[deprecated]], and
- using the -Werror=deprecated-declarations errors to rewrite calls to
  the deprecated raw_collect to calls to deprecated_raw_collect.

The new situation looks like (included here because the patch is large):
...
  void raw_collect (int regnum, gdb::array_view<gdb_byte> dst) const override;

  /* Deprecated overload of the above.  */
  [[deprecated]]
  void raw_collect (int regnum, void *dst) const;

  void deprecated_raw_collect (int regnum, void *dst) const
  {
DIAGNOSTIC_PUSH
DIAGNOSTIC_IGNORE_DEPRECATED_DECLARATIONS
    raw_collect (regnum, dst);
DIAGNOSTIC_POP
  }
...

There should be no functional changes.

There can be -Werror=deprecated-declarations build failures in files that are
not included in the --enable-targets=all configuration.

Tested on aarch64-linux.