[gdb/build] Build extra nat targets as unit test
When refactoring, it often happens that files for many, if not all targets are
changed.  Testing all of these targets can be impractical.

The configure option --enable-targets=all allows us to build a fair amount of
files, but if you're developing on x86_64-linux then say sparc-nat.c or
sparc64-nat.c is not build.

Surprisingly, this just works:
...
$ ( cd build/gdb; make sparc-nat.o sparc64-nat.o )
  CXX    sparc-nat.o
  CXX    sparc64-nat.o
$
...

These files are not build in this configuration because:
- the platform doesn't necessarily supply required header files with the
  correct contens, and
- they don't add required functionality to gdb.

But we can build these as some form of unit test, where the test is
that the file can be build and linked into gdb.

Trying the same with gdb/sparc-linux-nat.c and gdb/sparc64-linux-nat.c does
result in some build errors, but it's easy to work around those:
- pass -DEXTRA_NAT if not build as primary native target, to be able to
  detect this use case,
- conditionally rename supply_gregset et al not to conflict with the primary
  native target, and
- don't register the target, again not to conflict with the primary native
  target.

The same works for gdb/ia64-linux-nat.c, but asm/ptrace_offsets.h is
unavailabe so we skip uses where there is no impact, and mock up PT_DBR
and __NR_getunwind to address the remaining uses.

Likewise for gdb/alpha-linux-nat.c, where alpha/ptrace.h is missing and we
mock up PC, GPR_BASE, and FPR_BASE.

The gdb/hppa-linux-nat.c case is the most complex one.  The typedefs brought
in via ucontext.h and sys/procfs.h are a mismatch with the code, so we use
local overrides.

Building on aarch64-linux and arm-linux revealed a few similar problems, and
on arm-linux also some 32-bit vs 64-bit issues in ia64-linux-nat.c.

Building these extra objects is enabled when both these configure flags are
active:
- --enable-targets=all
- --enable-unit-tests
and we're not building a release branch.

There are more files to do, but I figured I get some feedback at this point.

Is this a good idea?

In particular, do you expect this to reduce or increase maintenance overhead?

Any other comments?

Tested on x86_64-linux and aarch64-linux.

Build on arm-linux, both without and with --enable-64-bit-bfd.