[gdb/contrib] Add spellcheck-inc.sh
With spellcheck.sh, I run into:
...
$ ./gdb/contrib/spellcheck.sh gdb
TODO: onot -> note, not
...
because of this occurrence in gdb/ada-lang.c:
...
  {"Oabs", "\"abs\"", UNOP_ABS},
  {"Onot", "\"not\"", UNOP_LOGICAL_NOT},
  {"Oadd", "\"+\"", UNOP_PLUS},
...

Obviously this is intentional, so we need a mechanism to ignore this.

IWBN to have a fine-grained mechanism similar to spellcheck directives, but
it's good to have a per-file mechanism regardless, so for now add this and use
it.

Implementation-wise, it takes the shape of bash function ignore_word_in_file,
which is put into a separate new include file spellcheck-inc.sh.

If we'd put ignore_word_in_file in spellcheck.sh, then a change in
ignore_word_in_file would invalidate .git/spell-check.pat1.$md5sum, causing it
to be regenerated, which is slow.

Verified with shellcheck.