[gdb/tdep] Fix gdb.base/store.exp on s390x
On s390x-linux, I get:
...
(gdb) print l^M
$29 = 0^M
(gdb) FAIL: gdb.base/store.exp: var doublest l; print old l, expecting -1
...

So, we're in wack_doublest trying to print l, which is a copy of parameter u:
...
  register doublest l = u, r = v;
...
which does have the expected value:
...
(gdb) p u
$1 = -1
...
which is a long double, 16 bytes and looks like this:
...
(gdb) p /x u
$3 = 0xbfff0000000000000000000000000000
...

Parameter u is passed in two registers:
...
 <2><6a5>: Abbrev Number: 15 (DW_TAG_formal_parameter)
    <6a6>   DW_AT_name        : v
    <69e>   DW_AT_location    : 6 byte block: 50 93 8 51 93 8 \
      (DW_OP_reg0 (r0); DW_OP_piece: 8; DW_OP_reg1 (r1); DW_OP_piece: 8)
...
and indeed we find the msw in r0 and the lsw in r1:
...
(gdb) p /x $r0
$4 = 0xbfff000000000000
(gdb) p /x $r1
$5 = 0x0
(gdb)
...

Likewise, variable l consists of two registers:
...
 <2><6b5>: Abbrev Number: 13 (DW_TAG_variable)
    <6b6>   DW_AT_name        : l
    <6be>   DW_AT_location    : 6 byte block: 68 93 8 69 93 8 \
      (DW_OP_reg24 (f8); DW_OP_piece: 8; DW_OP_reg25 (f10); DW_OP_piece: 8)
...
and we find the same values there:
...
(gdb) p /x $f8
$6 = 0xbfff000000000000
(gdb) p /x $f10
$7 = 0x0
...

So, we get the expected results when fetching the value from two gprs, but not
when fetching the value from two fprs.

The root cause for this seems to be this bit in s390_dwarf_reg_to_regnum:
...
  if (tdep->v0_full_regnum == -1)
    {
      ...
    }
  else
    {
      if (gdb_reg >= S390_F0_REGNUM && gdb_reg <= S390_F15_REGNUM)
       gdb_reg = gdb_reg - S390_F0_REGNUM + tdep->v0_full_regnum;
    }
...

So, instead of fetching the values of f8 and f10, we fetch the values of v8
and v10, which look like this:
...
(gdb) p /x $v8
$10 = {v4_float = {0xbfff0000, 0x0, 0x0, 0x0}, v2_double = {0xbfff000000000000, 0x0},
  v16_int8 = {0xbf, 0xff, 0x0 <repeats 14 times>}, v8_int16 = {0xbfff, 0x0, 0x0, 0x0, 0x0,
    0x0, 0x0, 0x0}, v4_int32 = {0xbfff0000, 0x0, 0x0, 0x0}, v2_int64 = {0xbfff000000000000,
    0x0}, uint128 = 0xbfff0000000000000000000000000000}
(gdb) p /x $v10
$11 = {v4_float = {0x0, 0x0, 0x0, 0x0}, v2_double = {0x0, 0x0}, v16_int8 = {
    0x0 <repeats 16 times>}, v8_int16 = {0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0}, v4_int32 = {
    0x0, 0x0, 0x0, 0x0}, v2_int64 = {0x0, 0x0}, uint128 = 0x0}
...

Because v8 is 16 bytes and we fetch only 8 bytes, this bit in rw_pieced_value is
activated:
...
		/* Big-endian, and we want less than full size.  */
		bits_to_skip += reg_bits - (p->offset + p->size);
...
and bits_to_skip gets set to 64, which gets us the value 0x0 instead of
0xbfff000000000000.

So, is there a bug in rw_pieced_value?

The DWARF standard says about DW_OP_piece:
...
If the piece is located in a register, but does not occupy the entire
register, the placement of the piece within that register is defined by the
ABI.
...

The interpretation seems to be to get the lsw of the register, possibly
because this is documented for DW_OP_bit_piece:
...
If the location is a register, the offset is from the least
significant bit end of the register.
...

I don't understand this area well enough to decide whether there's a bug.

Fix this by simply avoiding the unnecessary conversion from f0 to v0 in
s390_dwarf_reg_to_regnum.

Tested on s390x-linux.

Fixes test-cases:
- gdb.ada/O2_float_param.exp
- gdb.base/savedregs.exp
- gdb.base/store.exp