gdb: block SIGTERM during fetch_inferior_event
I'm posting this as RFC.  I started looking at this when I got a CI
failure email from Linaro about a regression on
gdb.base/gdb-sigerm.exp on ARM.  Turns out is wasn't my fault, it's
just the precise failure point moves about, so it can look like a
regression.

After looking at it for a bit I realised this was PR gdb/31061, and
there have already been attempts to fix this over the last few years.

What I have here is different than any of the previous approaches
posted, but I'm still not entirely sure this is the right solution,
but I thought I'd share it.  Might be nice to see if we can get this
fixed.

The patch might still need some cleanup, but it should be good enough
to discuss this approach.

Thanks,
Andrew