Backport i386_canonicalize_syscall rewrite to gdb-16-branch
Commit fbfb29b304e ("[gdb/tdep] Rewrite i386_canonicalize_syscall") fixes
PR32770, which reproduces on the gdb-16-branch, but the commit is not ideal
for backporting because it completely rewrites i386_canonicalize_syscall.

Instead, this is a version of the patch that adds a single line entry for each
syscall value for which i386_canonicalize_syscall gives a different result
with and without the patch.

Consequently, the two versions give identical results.  I've checked this for
syscalls 0 to 466.

Tested on x86_64-linux with target board unix/-m32, on top of gdb-16-branch.

PR tdep/32770
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=32770